Override implementation of authenticate.
Update authenticate method to include the original req parameter so that other request query parameters can be extracted and added to the request for an OAuth token.

Onshape requires a company_id parameter (if available) on the OAuth requests in order to pair a users token with the users company.